### 2023-09-10 #5829 @arfon
@PetrKorab – I don't think @editorialbot knows how to [find your paper](https://github.com/PetrKorab/Arabica/blob/main/paper%20joss/paper.md) when it's in a path with a space in it. Could you please rename the `paper joss` folder to `paper-joss`?
