### 2023-09-10 #5829 @arfon
Also, it would be very helpful to describe how your submission varies in scope/goals to this one: https://github.com/openjournals/joss-reviews/issues/4886
