### 2023-09-18 fd4ff2b
Delete paper-joss directory
