+++ paper-joss/paper.bib
@@ -1,138 +1,0 @@
-@article{Balduzzi:2023,
-  	author = {{Balduzzi}, S. and {Rücker}, G. and {Nikolakopoulou}, A. and {Papakonstantinou}, T. and {Salanti}, G. and {Efthimiou}, O. and {Schwarzer}, G.},
-    title = {netmeta: An R Package for Network Meta-Analysis Using Frequentist Methods},
-  	journal = {Journal of Statistical Software},
-  	keywords = {Network Meta-Analysis, R, Inconsistency, Component Network Meta-Analysis},
-    year = 2023,
-    volume = 558,
-  	month = mar,
-    doi = {10.18637/jss.v106.i02},
-    url = {https://www.jstatsoft.org/article/view/v106i02},
-}
-
-@misc{cleantext:2021,
-    author = {{Guidiva}, P.},
-    title = {Cleantext—An open-source python package to clean raw text data},
-    year = {2021},
-    publisher = {PyPI},
-    journal = {Python Package Index},
-    url = {https://pypi.org/project/cleantext}
-}
-
-@misc{FinVADER:2023,
-    author = {{Koráb}, P.},
-    title = {FinVADER: VADER sentiment classifier updated with financial lexicons},
-    year = {2023},
-    publisher = {PyPI},
-    journal = {Python Package Index},
-    url = {https://pypi.org/project/finvader}
-}
-
-@article{Hunter:2007,
-  	author = {{Hunter}, J.~D.},
-    title = {Matplotlib: A 2D Graphics Environment},
-  	journal = {Computing in Science & Engineering},
-  	keywords = {Python, scripting languages, application development, scientific programming},
-    year = 2007,
-    volume = 9,
-  	month = mar,
-    doi = {10.1109/MCSE.2007.55},
-    url = {https://ieeexplore.ieee.org/document/4160265},
-}
-
-@article{Hutto:2014,
-    author = {{Hutto}, C. and {Gilbert}, E.},
-    title = {VADER: A Parsimonious Rule-Based Model for Sentiment Analysis of Social Media Text},
-    journal = {Proceedings of the International AAAI Conference on Web and Social Media},
-    keywords = {Sentiment Analysis, Social Media, Validated Sentiment Lexicon, Human Centered Computing},
-    year = 2014,
-    volume = 8,
-    month = jan,
-    doi = {10.1609/icwsm.v8i1.14550},
-    url = {https://ojs.aaai.org/index.php/ICWSM/article/view/14550},
-}
-
-@book{Jenks:1977,
-    url = {https://openlibrary.org/books/OL22018775M/Optimal_data_classification_for_choropleth_maps},
-    Author = {{Jenks}, G.~F..},
-    Booktitle = {Optimal data classification for choropleth maps, by  George F. Jenks. Published by University of Kansas},
-    Publisher = {University of Kansas},
-    Title = {Optimal data classification for choropleth maps},
-    Year = 1977
-}
-
-@misc{jenkspy:2023,
-    author = {{Viry}, T.},
-    title = {Jenkspy- Compute Natural Breaks (Fisher-Jenks algorithm)},
-    year = {2023},
-    publisher = {PyPI},
-    journal = {Python Package Index},
-    url = {https://pypi.org/project/jenkspy}
-}
-
-@article{Loper:2002,
-    url = {https://arxiv.org/abs/cs/0205028},
-    Archiveprefix = {arXiv},
-    Author = {{Loper}, E. and {Bird}, S.},
-    Journal = {ArXiv e-prints},
-    Month = may,
-    Title = {NLTK: The Natural Language Toolkit},
-    Year = 2002
-}
-
-@book{McKinney:2013,
[…]
