+++ paper.bib
@@ -110,6 +110,15 @@
     url = {https://ideas.repec.org/c/boc/bocode/s458319.html}
 }
 
+@misc{pyarabic:2022,
+    author = {{Zerrouki}, T.},
+    title = {PyArabic: A Python package for Arabic text.},
+    year = {2022},
+    publisher = {PyPI},
+    journal = {Python Package Index},
+    url = {https://pypi.org/project/PyArabic}
+}
+
 @misc{texthero:2021,
     author = {{Guidiva}, P.},
     title = {texthero — Text preprocessing, representation and visualization from zero to hero.},
