+++ paper.bib
@@ -19,7 +19,7 @@
     url = {https://pypi.org/project/cleantext}
 }
 
-@misc{FinVADER:2023,
+@misc{finvader:2023,
     author = {{Koráb}, P.},
     title = {FinVADER: VADER sentiment classifier updated with financial lexicons},
     year = {2023},
@@ -101,7 +101,7 @@
     url = {https://www.frontiersin.org/articles/10.3389/fphys.2019.00203/full},
 }
 
-@misc{NETWORK:2017,
+@misc{network:2017,
     author = {{White}, I.},
     title = {NETWORK: Stata module to perform network meta-analysis},
     year = {2017},
