+++ paper.bib
@@ -1,4 +1,4 @@
-@article{Balduzzi:2023,
+omi@article{Balduzzi:2023,
   	author = {{Balduzzi}, S. and {Rücker}, G. and {Nikolakopoulou}, A. and {Papakonstantinou}, T. and {Salanti}, G. and {Efthimiou}, O. and {Schwarzer}, G.},
     title = {netmeta: An R Package for Network Meta-Analysis Using Frequentist Methods},
   	journal = {Journal of Statistical Software},
@@ -120,7 +120,7 @@
 }
 
 @misc{texthero:2021,
-    author = {{Guidiva}, P.},
+    author = {{Besomi}, J.},
     title = {texthero — Text preprocessing, representation and visualization from zero to hero.},
     year = {2021},
     publisher = {PyPI},
