### 2023-09-18 #5829 @PetrKorab
@arfon Thanks. I moved the paper files to the main directory (https://github.com/PetrKorab/Arabica). The paper should now compile without problems. I also updated the paper with the contribution to the literature you mentioned.
