+++ docs/source/index.rst
@@ -6,7 +6,7 @@
 Welcome to Arabica's documentation!
 ===================================
 **Arabica** is a python library for exploratory data analysis specifically designed for time-series text data.
-It reflects the reality that many text datasets are collected as repeated observations over time (social media conversation, 
+It reflects the reality that many text datasets are collected as repeated observations over time (social media conversations, 
 research metadata, product reviews, newspaper headlines, central bankers' communication, etc.).
 
 * **Descriptive n-gram analysis**: n-gram frequencies
