+++ docs/source/index.rst
@@ -44,4 +44,4 @@ Contents
 ------
 
 
-I have created this project in my free time, and I hope Arabica will save you some time. You can `invite me for coffe <https://www.buymeacoffee.com/PetrKorab>`_ if Arabica helps you with your business project, thesis, or research paper.
+I have created this project in my free time, and I hope Arabica will save you some time. You can `invite me for coffe <https://www.buymeacoffee.com/PetrKorab>`_ if Arabica helps you with your project, thesis, or research paper.
