### 2023-09-28 #5829 @arfon
@PetrKorab – due to the relatively limited size of this submission, I'm going to ask the JOSS editorial team to take a look at whether this submission is in scope for us. This will likely take a couple of weeks to complete.
