### 2023-10-02 #5829 @arfon
@PetrKorab – I don't see any way of testing this software. Could you please explain to me how this library is tested?
