+++ .travis.yml
@@ -3,7 +3,8 @@ language: python
 python:
   - "3.8"
   - "3.10"
+  - "pypy3"  
 install:
-  - pip install -r requirements.txt
+  - pip install
 script:
   - pytest
