+++ .travis.yml
@@ -5,6 +5,6 @@ python:
   - "3.10"
   - "pypy3"  
 install:
-  - pip install
+  - pip install arabica
 script:
   - pytest
