+++ .travis.yml
@@ -1,10 +1,0 @@
-
-language: python
-python:
-  - "3.8"
-  - "3.10"
-  - "pypy3"  
-install:
-  - pip install arabica
-script:
-  - pytest
