+++ .travis.yml
@@ -1,0 +1,21 @@
+
+language: python
+python:
+  - "3.8"
+  - "3.10"
+install:
+  - pip install pandas==1.4.0
+                nltk==3.6.2
+                regex==2022.10.31
+                finvader
+                matplotlib==3.6.0
+                matplotlib-inline==0.1.6
+                plotnine==0.10.1
+                wordcloud==1.8.2.2
+                jenkspy==0.3.2
+                vaderSentiment==3.3.2
+                cleantext==1.1.4
+                mizani==0.9.2
+                pillow==9.4.0
+script:
+  - pytest
