+++ .travis.yml
@@ -1,0 +1,8 @@
+language: python
+python:
+  - "3.8"
+  - "3.10"
+install:
+  - pip install -r requirements.txt
+script:
+  - python -m unittest discover
