+++ .travis.yml
@@ -1,11 +1,0 @@
-language: python
-python:
-  - "3.8"
-  - "3.10"
-install:
-  - pip install -r requirements.txt
-
-jobs:
-  include:
-  - stage: "App Tests"
-    script: docker-compose
