+++ .travis.yml
@@ -4,5 +4,4 @@ python:
   - "3.10"
 install:
   - pip install -r requirements.txt
-
 script: pytest
