+++ requirements.txt
@@ -10,4 +10,4 @@ jenkspy==0.3.2
 vaderSentiment==3.3.2
 cleantext==1.1.4
 mizani==0.9.2
-pillow == 9.4.0
+pillow==9.4.0
