+++ .travis.yml
@@ -16,4 +16,3 @@ install:
                 cleantext==1.1.4
                 mizani==0.9.2
                 pillow==9.4.0
-script: pytest
