+++ arabica/__init__.py
@@ -1,6 +1,0 @@
-__version__ = "1.4.6"
-
-
-from .arabica_freq import *
-from .cappuccino import *
-from .coffee_break import *
