+++ arabica/arabica_freq.py
@@ -1,234 +1,0 @@
-"""arabica_freq method"""
-
-
-import pandas as pd
-import numpy as np
-import nltk
-from nltk.util import ngrams
-from nltk.util import bigrams as bi
-from nltk.util import trigrams as tri
-
-from .preprocess import *
-from .clean_numbers import remove_numbers
-from .stopwords import remove_stopwords
-from .group import grouper
-from .clean_ngram import *
-
-
-def arabica_freq(text: str,                 # Text
-                 time: str,                 # Time
-                 date_format: str,          # Date format: 'eur' - European, 'us' - American
-                 stopwords: [],             # Languages for stop words
-                 skip: [],                  # Remove additional strings
-                 time_freq: str = '',       # Aggregation period: 'Y'/'M'/'D', if no aggregation: 'ungroup'
-                 max_words: int = '',       # Max number for most frequent n-grams displayed for each period
-                 numbers: bool = False,     # Remove all digits
-                 lower_case: bool = False   # Lowercase text before cleaning and frequency analysis
-                 ):
-
-    data=pd.DataFrame()
-    data['text']=text
-
-    if time_freq != 'ungroup':
-        data['time']=time
-    else:
-        data['time']='2016-08-30'
-
-    if date_format == 'eur':
-        data['time']=data['time'].astype(str)
-        data['time'] = pd.to_datetime(data['time'], dayfirst=True)
-
-    elif date_format == 'us':
-        data['time']=data['time'].astype(str)
-        data['time'] = pd.to_datetime(data['time'], dayfirst=False)
-
-    data=data.set_index(data['time'])
-    data.replace("", float("NaN"), inplace=True)
-    data.dropna(subset = ["text"], inplace=True)
-
-    if skip is not None:
-        data['text'] = data.text.str.replace('|'.join(skip), '.', regex=True).str.strip()
-
-    if stopwords is not None:
-        data['text'] = data.text.map(lambda x: remove_stopwords(x,stopwords=[stopwords]))
-
-    if lower_case is True:
-        data['text'] = np.vectorize(lower_casing)(data['text'])
-
-    data['text'] = np.vectorize(preprocess)(data['text'])
-
-    if numbers is True:
-        data['text'] = np.vectorize(remove_numbers)(data['text'])
-
-    periods = []
-    unigrams_freq = []
-    bigrams_freq = []
-    trigrams_freq = []
-
-    for i, row in data.iterrows():
-        period = row[1]
-        periods.append(period)
-        word = row[0]
-        token = str(word).split()
-        unigrams = list(ngrams(token, 1))
-        bigrams = list(bi(token))
-        trigrams = list(tri(token))
-        unigrams_freq.append(unigrams)
-        bigrams_freq.append(bigrams)
-        trigrams_freq.append(trigrams)
-
-    df = pd.DataFrame()
-    df['period'] = periods
-    df['period'] = df['period'].astype(str)
-    df['period'] = pd.to_datetime(df['period'])
-    df = df.set_index(df['period'])
-
-    df['unigrams'] = unigrams_freq
-    df['bigrams'] = bigrams_freq
-    df['trigrams'] = trigrams_freq
-
-    if time_freq == 'ungroup':
-
-        df['unigrams'] = np.vectorize(ungroup_cleaning)(df['unigrams'])
-
[…]
