### 2023-10-07 86811e3
Delete arabica/cappuccino.py
