+++ arabica/cappuccino.py
@@ -1,462 +1,0 @@
-# Data visualization module
-
-from plotnine import *
-from arabica import arabica_freq
-from wordcloud import WordCloud
-import matplotlib.pyplot as plt
-from matplotlib.pyplot import figure
-import pandas as pd
-import numpy as np
-
-
-def cappuccino(text: str,                  # Text
-               time: str,                  # Time
-               date_format: str,           # Date format: 'eur' - European, 'us' - American
-               stopwords: [],              # Languages for stop words
-               skip: [ ],                  # Remove additional strings
-               plot: str = '',             # Chart type: 'wordcloud'/'heatmap'/'line'
-               ngram: int = '',            # N-gram size, 1 = unigram, 2 = bigram, 3 = trigram
-               time_freq: str = '',        # Aggregation period: 'Y'/'M', if no aggregation: 'ungroup'
-               max_words: int = '',        # Max number for most frequent n-grams displayed for each period
-               numbers: bool = False,      # Remove numbers
-               lower_case: bool = False):  # Lowercase text before cleaning and frequency analysis
-
-
-    result = arabica_freq(text = text,
-                          time = time,
-                          date_format = date_format,
-                          time_freq = time_freq,
-                          max_words = max_words,
-                          stopwords = stopwords,
-                          skip = skip,
-                          numbers = numbers,
-                          lower_case = lower_case)
-
-
-    if ngram == 1:
-        if time_freq == 'ungroup':
-            if plot == 'wordcloud':
-                unigrams_df = result[['unigram','unigram_freq']]
-                unigrams_df = unigrams_df.dropna()
-                unigrams_df['unigram_freq'] = unigrams_df['unigram_freq'].astype(int)
-                unigrams_dict = dict(unigrams_df.values)
-                fig = figure()
-                wordcloud_unigram = WordCloud(background_color="white",
-                                              colormap='tab20b',
-                                              max_font_size=250,
-                                              width=2000,
-                                              height=1200,
-                                              margin=10).generate_from_frequencies(unigrams_dict)
-                plt.figure(dpi=1200)
-                plt.imshow(wordcloud_unigram,interpolation="bilinear")
-                fig.set_size_inches(10, 6, forward=True)
-                plt.axis('off')
-                picture = plt.show()
-
-
-        elif time_freq == 'M':
-            period = result['period']
-            subset = result['unigram']
-            subset = subset.dropna()
-            subset_expand = subset.str.split(pat=",", expand=True)
-            subset_expand.columns = ['unigram' + str(i + 1) for i in range(len(subset_expand.columns))]
-            subset_expand["id"] = subset_expand.index
-            test = subset_expand.merge(period, left_index = True, right_index = True)
-            reshaped = pd.wide_to_long(test, ['unigram'], i="id", j="period")
-            period_re = reshaped['period']
-            period_re.reset_index(drop=True, inplace=True)
-            reshaped_unigram = reshaped['unigram']
-            reshaped_unigram.reset_index(drop=True, inplace=True)
-            reshaped_unigram = reshaped_unigram.str.split(pat=": ", expand=True)
-            reshaped_unigram = reshaped_unigram.rename(columns={0: 'unigram', 1: 'frequency'})
-            reshaped_unigram = reshaped_unigram.dropna()
-            reshaped_unigram['frequency'] = reshaped_unigram['frequency'].astype(int)
-            reshaped_unigram = reshaped_unigram.merge(period_re, left_index = True, right_index = True)
-
-            if plot == 'heatmap':
-                reshaped_unigram = reshaped['unigram']
[…]
