### 2023-10-07 671e2cd
Delete arabica/stopwords.py
