### 2023-10-07 660d564
Delete arabica/coffee_break.py
