### 2023-10-07 #5829 @PetrKorab
@arfon I have added "tests" folder that contains the Travis CI test of the code generating the graphs used in the paper. I can provide additional manual tests, if necessary.
