### 2023-10-08 #5829 @arfon
> @arfon I have added "tests" folder that contains the Travis CI test of the code generating the graphs used in the paper. I can provide additional manual tests, if necessary.

This is probably useful, but not what I'm looking for. As a part of the JOSS review process, your software needs to have tests for reviewers to exercise (see https://joss.readthedocs.io/en/latest/review_criteria.html#tests). Without these, the package cannot be reviewed.
