+++ .github/workflows/test_python38.yml
@@ -19,4 +19,4 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
     - name: Run script
-      run: python tests.py
+      run: python test.py
