+++ .circleci/config.yml
@@ -1,12 +1,0 @@
-version: 2.1
-jobs:
-  build:
-    docker:
-      - image: cimg/base:current
-    resource_class: petrkorab/arabica
-    steps:
-      - run: echo "Hi I'm on a Container Runner!"
-workflows:
-  build-workflow:
-    jobs:
-      - build
