+++ .circleci/main.yml
@@ -1,27 +1,0 @@
-version: 2.1
-parameters:
-  GHA_Event:
-    type: string
-    default: ""
-  GHA_Actor:
-    type: string
-    default: ""
-  GHA_Action:
-    type: string
-    default: ""
-  GHA_Meta:
-    type: string
-    default: ""
-    
-on:
-  release:
-    types: [published]
-jobs:
-  trigger-circleci:
-    runs-on: ubuntu-latest
-    steps:
-      - name: <customize name>
-        id: <customize id>
-        uses: circleci/trigger_circleci_pipeline@v1.0
-        env:
-          CCI_TOKEN: $
