+++ .circleci/config.yml
@@ -25,6 +25,8 @@ workflows:
     jobs:
       - say-hello
             steps:
-      - add_ssh_keys:
-            fingerprints:
-              - "SHA256:mXV4fwR8uHei6T1tUI+kkKmyesvZoxT8B6OIa+DNIi0 email@example.com"
+      - deploy-job:
+            steps:
+              - add_ssh_keys:
+                  fingerprints:
+                    - "SHA256:mXV4fwR8uHei6T1tUI+kkKmyesvZoxT8B6OIa+DNIi0 email@example.com"
