+++ .circleci/config.yml
@@ -39,9 +39,3 @@ workflows:
   say-hello-workflow:
     jobs:
       - say-hello
-            steps:
-      - deploy-job:
-            steps:
-              - add_ssh_keys:
-                  fingerprints:
-                    - "SHA256:mXV4fwR8uHei6T1tUI+kkKmyesvZoxT8B6OIa+DNIi0 email@example.com"
