+++ .circleci/config.yml
@@ -6,12 +6,12 @@ jobs:
     steps:
       - checkout
       - restore_cache:
-          key: deps1-{{ .Branch }}-{{ checksum "requirements_TESTS.txt" }}
+          key: deps1-{{ .Branch }}-{{ checksum "requirements.txt" }}
       - run:
           command: |
             python3 -m venv venv
             . venv/bin/activate
-            pip install -r requirements_TESTS.txt
+            pip install -r requirements.txt
       - save_cache:
           key: deps1-{{ .Branch }}-{{ checksum "requirements.txt" }}
           paths:
@@ -20,11 +20,10 @@ jobs:
           name: Running tests
           command: |
             . venv/bin/activate
-            python3 test.py
+            python3 tests.py
       - store_artifacts:
           path: test-reports/
           destination: python_app
-
 
 # version: 2.1
 # parameters:
