+++ .circleci/config.yml
@@ -1,64 +1,0 @@
-version: 2.1
-jobs:
-  build:
-    docker:
-      - image: cimg/python:3.10.1
-    steps:
-      - checkout
-      - restore_cache:
-          key: deps1-{{ .Branch }}-{{ checksum "requirements_TEST.txt" }}
-      - run:
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            pip install -r requirements.txt
-      - save_cache:
-          key: deps1-{{ .Branch }}-{{ checksum "requirements_TEST.txt" }}
-          paths:
-            - "venv"
-      - run:
-          name: Running tests
-          command: |
-            . venv/bin/activate
-            python3 tests.py
-      - store_artifacts:
-          path: test-reports/
-          destination: python_app
-
-# version: 2.1
-# parameters:
-#   GHA_Actor:
-#     type: string
-#     default: ""
-#   GHA_Action:
-#     type: string
-#     default: ""
-#   GHA_Event:
-#     type: string
-#     default: ""
-#   GHA_Meta:
-#     type: string
-#     default: ""
-
-# # Define a job to be invoked later in a workflow.
-# # See: https://circleci.com/docs/configuration-reference/#jobs
-# jobs:
-#   say-hello:
-#     # Specify the execution environment. You can specify an image from Docker Hub or use one of our convenience images from CircleCI's Developer Hub.
-#     # See: https://circleci.com/docs/configuration-reference/#executor-job
-#     docker:
-#       - image: cimg/base:stable
-#     # Add steps to the job
-#     # See: https://circleci.com/docs/configuration-reference/#steps
-#     steps:
-#       - checkout
-#       - run:
-#           name: "Say hello"
-#           command: "echo Hello, World!"
-
-# # Orchestrate jobs using workflows
-# # See: https://circleci.com/docs/configuration-reference/#workflows
-# workflows:
-#   say-hello-workflow:
-#     jobs:
-#       - say-hello
