### 2023-10-15 e5b991e
Delete hello.py
