+++ .travis.yml
@@ -1,0 +1,21 @@
+language: python  # Adjust this based on your project's programming language
+
+branches:
+  only:
+    - main
+    - /^pull\//
+    
+
+# Specify the version(s) of your language you want to test with
+python:
+  - 3.8
+  - 3.9
+  # Add more versions if needed
+
+# Install dependencies
+install:
+  - pip install -r requirements.txt  # Adjust for your project
+
+# Run tests
+script:
+  - python tests.py  # Adjust for your project
