+++ tests/tests.py
@@ -16,7 +16,7 @@ cappuccino(text = data['title'],
            time = data['publicationYear'],
            date_format = 'us',                # Uses US-style date format to parse dates
            plot = 'wordcloud',
-           ngram = 2,                         # N-gram size, 1 = unigram, 2 = bigram, 3 = trigram
+           ngram = 1,                         # N-gram size, 1 = unigram, 2 = bigram, 3 = trigram
            time_freq = 'ungroup',             # No time aggregation
            max_words = 150,                   # Displays 150 most frequent bigrams
            stopwords = ['english'],           # Remove English stopwords
@@ -38,7 +38,7 @@ cappuccino(text = data['title'],
            plot = 'heatmap',
            ngram = 2,                         # N-gram size, 1 = unigram, 2 = bigram
            time_freq = 'Y',                   # Aggregation period, 'M' = monthly, 'Y' = yearly
-           max_words = 6,                     # Displays 6 most frequent bigrams for each period
+           max_words = 4,                     # Displays 6 most frequent bigrams for each period
            stopwords = ['english'],           # Remove English stopwords
            skip = None,                       # Remove additional stopwords
            numbers = True,                    # Remove numbers
@@ -55,7 +55,7 @@ cappuccino(text = data['title'],
            time = data['publicationYear'],
            date_format = 'us',                # Uses US-style date format to parse dates
            plot = 'line',
-           ngram = 2,                         # N-gram size, 1 = unigram, 2 = bigram
+           ngram = 1,                         # N-gram size, 1 = unigram, 2 = bigram
            time_freq = 'Y',                   # Aggregation period, 'M' = monthly, 'Y' = yearly
            max_words = 4,                     # Displays 6 most frequent bigrams for each period
            stopwords = ['english'],           # Remove English stopwords
@@ -79,6 +79,6 @@ coffee_break(text = data['title'],
              time_freq = 'Y',                  # Yearly aggregation
              preprocess = True,                # Clean data - digits and punctuation
              skip = None,                      # Remove additional stop words
-             n_breaks = 2)                     # Find 2 breakpoints
+             n_breaks = 3)                     # Find 2 breakpoints
 
 print("lineplot with sentiment and breakpoint analysis results generated, save manually as jpg./.png")
