### 2023-10-15 #5829 @PetrKorab
@arfon The "tests" folder in the main branch now contains automated tests with Travis CI (Circle Ci alike). The reviews can fork the repository and make pull requests to run the test.
