### 2023-10-22 e4c872e
Delete .github/workflows/test_python38.yml
