+++ docs/source/Use cases and tutorials.rst
@@ -4,4 +4,5 @@ Use cases and tutorials
 
 * **Business Intelligence:** Customer Satisfaction Measurement with N-gram and Sentiment Analysis  `(article here <https://towardsdatascience.com/customer-satisfaction-measurement-with-n-gram-and-sentiment-analysis-547e291c13a6?sk=62f9decb619744c96c49735ff09653c3>`_)   
 * **Research meta-data analysis:** Research Article Meta-data Description Made Quick and Easy `(article here <https://pub.towardsai.net/research-article-meta-data-description-made-quick-and-easy-57754e54b550?sk=82477c74a159855f211b09b53026dedc>`_)   
-* **Twitter tweets analysis**
+* **Media coverage text mining**
+* **Social media analysis**
