+++ README.md
@@ -30,7 +30,7 @@ It reads dates in:
 
 ## Installation
 
-Arabica requires Python 3.8 -3.10, [NLTK](http://www.nltk.org) - stop words removal,
+Arabica requires *Python 3.8 -3.10*, [NLTK](http://www.nltk.org) - stop words removal,
 [cleantext](https://pypi.org/project/cleantext/#description) - text cleaning, [wordcloud](https://pypi.org/project/wordcloud) - word cloud visualization,
 [plotnine](https://pypi.org/project/plotnine) - heatmaps and line graphs, [matplotlib](https://pypi.org/project/matplotlib/) - word clouds and graphical operations,
 [vaderSentiment](https://pypi.org/project/vaderSentiment) - sentiment analysis, [finvader](https://pypi.org/project/finvader) - financial sentiment analysis,
