### 2024-01-07 #6186 @editorialbot
[REVIEW]: Arabica: A Python package for exploratory analysis of text data
**Submitting author:** @PetrKorab (<a href="http://orcid.org/0000-0003-2513-8981">Petr Koráb</a>)
**Repository:** https://github.com/PetrKorab/Arabica
**Branch with paper.md** (empty if default branch): main
**Version:** 1.7.1
