+++ paper.bib
@@ -116,6 +116,7 @@ omi@article{Balduzzi:2023,
     year = {2022},
     publisher = {PyPI},
     journal = {Python Package Index},
+    doi = {10.21105/joss.04886},
     url = {https://pypi.org/project/PyArabic}
 }
 
