### 2024-01-08 #6186 @PetrKorab
@oliviaguest I've added DOI for PyArabic: A Python package for Arabic text.
