### 2024-01-26 #6186 @linuxscout
> @oliviaguest I've added DOI for PyArabic: A Python package for Arabic text.

I believe the reference to this package, PyArabic, is inappropriate for this work because PyArabic is a modest library for basic Arabic text functionality. I recommend citing more complex text processing frameworks or data visualizations.
