### 2024-01-26 PetrKorab/Arabica#11 @linuxscout
Automated tests:
requirement for JOSS journal.
Automated tests:
Are there automated tests or manual steps described so that the functionality of the software can be verified?
