### 2024-01-26 #6186 @linuxscout
@PetrKorab
I Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
I added an issue to repo: https://github.com/PetrKorab/Arabica/issues/11
