### 2024-01-28 PetrKorab/Arabica#11 @linuxscout
Thank you,
I think that the given script is an example, but not an automated tests,
I suggest to add more test examples using unittest.
