+++ .travis.yml
@@ -8,9 +8,9 @@ branches:
     - /^pull\//
     
 python:
+- "3.10"
   # - 3.8
   # - 3.9
-  - 3.10
 
 # Install dependencies
 install:
