+++ .travis.yml
@@ -8,7 +8,7 @@ branches:
     - /^pull\//
     
 python:
-- "3.10"
+- "3.10.0"
   # - 3.8
   # - 3.9
 
