+++ .travis.yml
@@ -8,6 +8,7 @@ branches:
     - /^pull\//
     
 python:
+ - "3.8"
  - "3.9"
  - "3.10"
 
