+++ .travis.yml
@@ -10,7 +10,6 @@ branches:
 python:
  - "3.8"
  - "3.9"
- - "3.10"
 
 # Install dependencies
 install:
@@ -19,4 +18,4 @@ install:
 # Run tests
 script:
   - python tests.py 
-  # - arabica_freq_unit_test.py
+  - arabica_freq_unit_test.py
