### 2024-02-04 bf8aacd
Update arabica_freq_unit_test.py
