+++ .travis.yml
@@ -9,7 +9,7 @@ branches:
     
 python:
  - "3.8"
- - "3.9"
+ # - "3.9"
 
 # Install dependencies
 install:
