+++ .travis.yml
@@ -19,4 +19,4 @@ install:
 script:
   - python paper_graphs.py 
   - python arabica_freq_unit_test.py
-  - python cappuccino_unit_test.pipy
+  - python cappuccino_unit_test.py
