+++ tests/coffee_break_unit_test.py
@@ -1,80 +1,118 @@
-## Test for visualization method: cappuccino
-## It relies on the n-gram frequency matrix from arabica_freq
-## n-gram frequencies are visualised with Matplotlib or Plotnine
-
-
 import unittest
 import pandas as pd
-from arabica import cappuccino
+import numpy as np
+from arabica import coffee_break
 
-class TestCappuccinoFunction(unittest.TestCase):
+class TestCoffeeBreakFunction(unittest.TestCase):
 
     data = pd.DataFrame({
-        'text': ["You may find yourself trying to decide between comparable crystallized ginger offerings from Reeds",
-                 "Reed’s has a lovely raw cane sugar flavor, and is sweeter and more mellow than The Ginger People’s."],
-        'time': ["2016-08-30", "2016-08-31"]})
+        'text': ['Same folks said daikon paste could treat a cytokine storm #PfizerBioNTech https://t.co/xeHhIMg1kF',
+                '''While the world has been on the wrong side of history this year, hopefully, the biggest vaccination effort we've evâ€¦ https://t.co/dlCHrZjkhm''',
+                 "#coronavirus #SputnikV #AstraZeneca #PfizerBioNTech #Moderna #Covid_19 Russian vaccine is created to last 2-4 yearsâ€¦ https://t.co/ieYlCKBr8P",
+                "Facts are immutable, Senator, even when you're not ethically sturdy enough to acknowledge them. (1) You were born iâ€¦ https://t.co/jqgV18kch4",
+                "Explain to me again why we need a vaccine @BorisJohnson @MattHancock #whereareallthesickpeople #PfizerBioNTechâ€¦ https://t.co/KxbSRoBEHq",
+                "Does anyone have any useful advice/guidance for whether the COVID vaccine is safe whilst breastfeeding?â€¦ https://t.co/EifsyQoeKN",
+                "it is a bit sad to claim the fame for success of #vaccination on patriotic competition between USA, Canada, UK andâ€¦ https://t.co/IfMrAyGyTP",
+                "There have not been many bright days in 2020 but here are some of the best",
+                "#BidenHarris winning #Election2020â€¦ https://t.co/77u4f8XXfx",
+                'Covid vaccine; You getting it?',            
+                '#CovidVaccine #covid19 #PfizerBioNTech #Moderna',
+                '#CovidVaccine',            
+                'States will start getting #COVID19Vaccine Monday, #US says',
+                '#pakustv #NYC #Healthcare #GlobalGoalsâ€¦ https://t.co/MksOvBvs5w',
+                "while deaths are closing in on the 300,000 mark... millions of people wait #PfizerBioNTech #Vaccine !",
+                'The first Uâ€¦ https://t.co/uQ3A2f7SVP',
+                '@cnnbrk #COVID19 #CovidVaccine #vaccine #Corona #PfizerBioNTech #bbcnews #NYTimes #BBCNews Best wishes to the USâ€¦ https://t.co/9xWmfU3LZj',
+                'The agency also released new information for health care providers and for patients as the US shipped millions of dâ€¦ https://t.co/sG6BtD7jD9',
+                'For all the women and healthcare providers who have been asking about the safety of the #PfizerBioNTechâ€¦ https://t.co/ow0Pglkwte',
+                '"Expect 145 sites across all the states to receive vaccine on Monday, another 425 sites on Tuesday," said the officâ€¦ https://t.co/HiUVYJzOBY',
+                "Trump announces #vaccine rollout 'in less than 24 hours",
+                "The first Americans will be vaccinated againstâ€¦ https://t.co/2FzQSMnhoY",
+                "UPDATED: #YellowFever &amp; #COVID19 #ImmunityPassports - Part Two",            
[…]
