### 2024-02-04 2712c9e
Update arabica_freq_unit_test.py
