+++ .travis.yml
@@ -9,7 +9,7 @@ branches:
     
 python:
  - "3.8"
- # - "3.9"
+ - "3.9"
 
 # Install dependencies
 install:
@@ -21,4 +21,3 @@ script:
   - python arabica_freq_unit_test.py
   - python cappuccino_unit_test.py
   - python coffee_break_unit_test.py
-  
