+++ tests/coffee_break_unit_test.py
@@ -18,11 +18,11 @@ class TestCoffeeBreakFunction(unittest.TestCase):
                 "#BidenHarris winning #Election2020â€¦ https://t.co/77u4f8XXfx",
                 'Covid vaccine; You getting it?',            
                 '#CovidVaccine #covid19 #PfizerBioNTech #Moderna',
-                '#CovidVaccine',            
+                #'#CovidVaccine',            
                 'States will start getting #COVID19Vaccine Monday, #US says',
                 '#pakustv #NYC #Healthcare #GlobalGoalsâ€¦ https://t.co/MksOvBvs5w',
                 "while deaths are closing in on the 300,000 mark... millions of people wait #PfizerBioNTech #Vaccine !",
-                'The first Uâ€¦ https://t.co/uQ3A2f7SVP',
+                # 'The first Uâ€¦ https://t.co/uQ3A2f7SVP',
                 '@cnnbrk #COVID19 #CovidVaccine #vaccine #Corona #PfizerBioNTech #bbcnews #NYTimes #BBCNews Best wishes to the USâ€¦ https://t.co/9xWmfU3LZj',
                 'The agency also released new information for health care providers and for patients as the US shipped millions of dâ€¦ https://t.co/sG6BtD7jD9',
                 'For all the women and healthcare providers who have been asking about the safety of the #PfizerBioNTechâ€¦ https://t.co/ow0Pglkwte',
@@ -51,8 +51,8 @@ class TestCoffeeBreakFunction(unittest.TestCase):
         'time': ['21/09/2009 15:27',
                 '21/09/2009 15:27',
                 '25/06/2020 23:30',
-                '26/01/2020 21:43',
-                '23/04/2020 17:58',
+                #'26/01/2020 21:43',
+                #'23/04/2020 17:58',
                 '26/01/2020 21:43',
                 '26/01/2020 21:43',
                 '25/03/2019 04:14',
