+++ .travis.yml
@@ -8,7 +8,7 @@ branches:
     - /^pull\//
     
 python:
- # - "3.8"
+ - "3.8"
  - "3.9"
 
 # Install dependencies
@@ -17,7 +17,7 @@ install:
 
 # Run tests
 script:
-  #- python paper_graphs.py 
-  #- python arabica_freq_unit_test.py
-  #- python cappuccino_unit_test.py
+  - python paper_graphs.py 
+  - python arabica_freq_unit_test.py
+  - python cappuccino_unit_test.py
   - python coffee_break_unit_test.py
