### 2024-02-04 1cb9200
Update coffee_break_unit_test.py
