+++ tests/coffee_break_unit_test.py
@@ -2,92 +2,93 @@
 
 import unittest
 import pandas as pd
+import numpy as np
 from arabica import coffee_break
 
-class TestCoffeeBreakFunction(unittest.TestCase):    
+class TestCoffeeBreakFunction(unittest.TestCase):
     # test data
     data = pd.DataFrame({
         'text': ['Same folks said daikon paste could treat a cytokine storm #PfizerBioNTech https://t.co/xeHhIMg1kF',
-                '''While the world has been on the wrong side of history this year, hopefully, the biggest vaccination effort we've evâ€¦ https://t.co/dlCHrZjkhm''',
-                 "#coronavirus #SputnikV #AstraZeneca #PfizerBioNTech #Moderna #Covid_19 Russian vaccine is created to last 2-4 yearsâ€¦ https://t.co/ieYlCKBr8P",
-                "Facts are immutable, Senator, even when you're not ethically sturdy enough to acknowledge them. (1) You were born iâ€¦ https://t.co/jqgV18kch4",
-                "Explain to me again why we need a vaccine @BorisJohnson @MattHancock #whereareallthesickpeople #PfizerBioNTechâ€¦ https://t.co/KxbSRoBEHq",
-                "Does anyone have any useful advice/guidance for whether the COVID vaccine is safe whilst breastfeeding?â€¦ https://t.co/EifsyQoeKN",
-                "it is a bit sad to claim the fame for success of #vaccination on patriotic competition between USA, Canada, UK andâ€¦ https://t.co/IfMrAyGyTP",
-                "There have not been many bright days in 2020 but here are some of the best",
-                "#BidenHarris winning #Election2020â€¦ https://t.co/77u4f8XXfx",
-                'Covid vaccine; You getting it?',            
-                '#CovidVaccine #covid19 #PfizerBioNTech #Moderna',
-                '#CovidVaccine',            
-                'States will start getting #COVID19Vaccine Monday, #US says',
-                '#pakustv #NYC #Healthcare #GlobalGoalsâ€¦ https://t.co/MksOvBvs5w',
-                "while deaths are closing in on the 300,000 mark... millions of people wait #PfizerBioNTech #Vaccine !",
-                'The first Uâ€¦ https://t.co/uQ3A2f7SVP',
-                '@cnnbrk #COVID19 #CovidVaccine #vaccine #Corona #PfizerBioNTech #bbcnews #NYTimes #BBCNews Best wishes to the USâ€¦ https://t.co/9xWmfU3LZj',
-                'The agency also released new information for health care providers and for patients as the US shipped millions of dâ€¦ https://t.co/sG6BtD7jD9',
-                'For all the women and healthcare providers who have been asking about the safety of the #PfizerBioNTechâ€¦ https://t.co/ow0Pglkwte',
-                '"Expect 145 sites across all the states to receive vaccine on Monday, another 425 sites on Tuesday," said the officâ€¦ https://t.co/HiUVYJzOBY',
-                "Trump announces #vaccine rollout 'in less than 24 hours",
-                "The first Americans will be vaccinated againstâ€¦ https://t.co/2FzQSMnhoY",
-                "UPDATED: #YellowFever &amp; #COVID19 #ImmunityPassports - Part Two",            
-                "#SARSCoV2 #PfizerBioNtech #Britain #Decemberâ€¦ https://t.co/qKT7Rst9aW",
-                'Coronavirus: Iran reports 8,201 new cases, 221 deaths in the last 24 hours #Iran #coronavirus #PfizerBioNTechâ€¦ https://t.co/mwDNAdmb7F',
-                '.@Pfizer will rake in billions from its expensive #CovidVaccine but @AlbertBourla refuses to cut the price for loweâ€¦ https://t.co/eC94w3TUl0',
[…]
