+++ tests/coffee_break_unit_test.py
@@ -9,11 +9,11 @@ class TestCoffeeBreakFunction(unittest.TestCase):
     # test data
     data = pd.DataFrame({
         'text': ['Same folks said daikon paste could treat a cytokine storm #PfizerBioNTech https://t.co/xeHhIMg1kF',
-                 # '''While the world has been on the wrong side of history this year, hopefully, the biggest vaccination effort we've evâ€¦ https://t.co/dlCHrZjkhm''',
-                 # "#coronavirus #SputnikV #AstraZeneca #PfizerBioNTech #Moderna #Covid_19 Russian vaccine is created to last 2-4 yearsâ€¦ https://t.co/ieYlCKBr8P",
-                 # "Facts are immutable, Senator, even when you're not ethically sturdy enough to acknowledge them. (1) You were born iâ€¦ https://t.co/jqgV18kch4",
-                 # "Explain to me again why we need a vaccine @BorisJohnson @MattHancock #whereareallthesickpeople #PfizerBioNTechâ€¦ https://t.co/KxbSRoBEHq",
-                 # "Does anyone have any useful advice/guidance for whether the COVID vaccine is safe whilst breastfeeding?â€¦ https://t.co/EifsyQoeKN",
+                 '''While the world has been on the wrong side of history this year, hopefully, the biggest vaccination effort we've evâ€¦ https://t.co/dlCHrZjkhm''',
+                 "#coronavirus #SputnikV #AstraZeneca #PfizerBioNTech #Moderna #Covid_19 Russian vaccine is created to last 2-4 yearsâ€¦ https://t.co/ieYlCKBr8P",
+                 "Facts are immutable, Senator, even when you're not ethically sturdy enough to acknowledge them. (1) You were born iâ€¦ https://t.co/jqgV18kch4",
+                 "Explain to me again why we need a vaccine @BorisJohnson @MattHancock #whereareallthesickpeople #PfizerBioNTechâ€¦ https://t.co/KxbSRoBEHq",
+                 "Does anyone have any useful advice/guidance for whether the COVID vaccine is safe whilst breastfeeding?â€¦ https://t.co/EifsyQoeKN",
                  # "it is a bit sad to claim the fame for success of #vaccination on patriotic competition between USA, Canada, UK andâ€¦ https://t.co/IfMrAyGyTP",
                  # "There have not been many bright days in 2020 but here are some of the best",
                  # "#BidenHarris winning #Election2020â€¦ https://t.co/77u4f8XXfx",
@@ -36,7 +36,7 @@ class TestCoffeeBreakFunction(unittest.TestCase):
                  # '.@Pfizer will rake in billions from its expensive #CovidVaccine but @AlbertBourla refuses to cut the price for loweâ€¦ https://t.co/eC94w3TUl0',
                  # 'The trump administration failed to deliver on vaccine promises, *shocker* #COVIDIOTS #coronavirus #CovidVaccineâ€¦ https://t.co/hew6eHTUrD',
                  # 'How much did the #fda get paid to approve this all of a sudden now? money makes the world go around. #vaccineâ€¦ https://t.co/GT9qVVNbKj',
-                 # 'Anyone wondering why day after #PfizerBioNTech approval in the UK people were getting vaccinated but all we are tolâ€¦ https://t.co/tPSyL9CUYE',
+                 'Anyone wondering why day after #PfizerBioNTech approval in the UK people were getting vaccinated but all we are tolâ€¦ https://t.co/tPSyL9CUYE',
                  "Trump announces #vaccine rollout 'in less than 24 hours",
                  "The first Americans will be vaccinated againstâ€¦ https://t.co/P9esXr3zpS",
[…]
