+++ tests/coffee_break_unit_test.py
@@ -14,28 +14,7 @@ class TestCoffeeBreakFunction(unittest.TestCase):
                  "Facts are immutable, Senator, even when you're not ethically sturdy enough to acknowledge them. (1) You were born iâ€¦ https://t.co/jqgV18kch4",
                  "Explain to me again why we need a vaccine @BorisJohnson @MattHancock #whereareallthesickpeople #PfizerBioNTechâ€¦ https://t.co/KxbSRoBEHq",
                  "Does anyone have any useful advice/guidance for whether the COVID vaccine is safe whilst breastfeeding?â€¦ https://t.co/EifsyQoeKN",
-                 # "it is a bit sad to claim the fame for success of #vaccination on patriotic competition between USA, Canada, UK andâ€¦ https://t.co/IfMrAyGyTP",
-                 # "There have not been many bright days in 2020 but here are some of the best",
-                 # "#BidenHarris winning #Election2020â€¦ https://t.co/77u4f8XXfx",
-                 # 'Covid vaccine; You getting it?',
-                 # '#CovidVaccine #covid19 #PfizerBioNTech #Moderna',
-                 # '#CovidVaccine',
-                 # 'States will start getting #COVID19Vaccine Monday, #US says',
-                 # '#pakustv #NYC #Healthcare #GlobalGoalsâ€¦ https://t.co/MksOvBvs5w',
-                 # "while deaths are closing in on the 300,000 mark... millions of people wait #PfizerBioNTech #Vaccine !",
-                 # 'The first Uâ€¦ https://t.co/uQ3A2f7SVP',
-                 # '@cnnbrk #COVID19 #CovidVaccine #vaccine #Corona #PfizerBioNTech #bbcnews #NYTimes #BBCNews Best wishes to the USâ€¦ https://t.co/9xWmfU3LZj',
-                 # 'The agency also released new information for health care providers and for patients as the US shipped millions of dâ€¦ https://t.co/sG6BtD7jD9',
-                 # 'For all the women and healthcare providers who have been asking about the safety of the #PfizerBioNTechâ€¦ https://t.co/ow0Pglkwte',
-                 # '"Expect 145 sites across all the states to receive vaccine on Monday, another 425 sites on Tuesday," said the officâ€¦ https://t.co/HiUVYJzOBY',
-                 # "Trump announces #vaccine rollout 'in less than 24 hours",
-                 # "The first Americans will be vaccinated againstâ€¦ https://t.co/2FzQSMnhoY",
-                 # "UPDATED: #YellowFever &amp; #COVID19 #ImmunityPassports - Part Two",
-                 # "#SARSCoV2 #PfizerBioNtech #Britain #Decemberâ€¦ https://t.co/qKT7Rst9aW",
-                 # 'Coronavirus: Iran reports 8,201 new cases, 221 deaths in the last 24 hours #Iran #coronavirus #PfizerBioNTechâ€¦ https://t.co/mwDNAdmb7F',
-                 # '.@Pfizer will rake in billions from its expensive #CovidVaccine but @AlbertBourla refuses to cut the price for loweâ€¦ https://t.co/eC94w3TUl0',
-                 # 'The trump administration failed to deliver on vaccine promises, *shocker* #COVIDIOTS #coronavirus #CovidVaccineâ€¦ https://t.co/hew6eHTUrD',
-                 # 'How much did the #fda get paid to approve this all of a sudden now? money makes the world go around. #vaccineâ€¦ https://t.co/GT9qVVNbKj',
+                 "ID19Vaccine Monday, #US says',                               
                  'Anyone wondering why day after #PfizerBioNTech approval in the UK people were getting vaccinated but all we are tolâ€¦ https://t.co/tPSyL9CUYE',
                  "Trump announces #vaccine rollout 'in less than 24 hours",
[…]
