### 2024-02-05 fc2916b
Update coffee_break_unit_test.py
