### 2024-02-05 8405576
Update coffee_break_unit_test.py
