### 2024-02-05 PetrKorab/Arabica#11 @PetrKorab
@linuxscout Thanks, I've also added automated tests for each of the main methods using unittest into the Tests folder.
