### 2024-02-05 #6186 @linuxscout
> @linuxscout @oliviaguest Thanks. Should I revise the paper now ?
>
> > > @oliviaguest I've added DOI for PyArabic: A Python package for Arabic text.
> > > @PetrKorab
> >
> >
> > I believe the reference to this package, PyArabic, is inappropriate for this work because PyArabic is a modest library for basic Arabic text functionality. I recommend citing more complex text processing frameworks or data visualizations.

automated tests are provided and the issue is  closed.
