### 2024-03-19 #6186 @amitkumarj441
I was able to run the python tests without any issue and most of the functionalities in the examples work well, though, I suggest checking the compatibility among the dependencies, such as the requirement for installing this should be at least Python 3.8.x. However, it is Python 3.9 in [setup.py](https://github.com/PetrKorab/Arabica/blob/24349afe523730bc0b04e5bfd02af6468174a843/setup.py#L19), I suggest fixing `python_requires=>3.9`. Also, I suggest an instruction for a source installation as `pip install Arabica` and given the version of this package upon installation is 1.7.7, you should fix it [here](https://github.com/PetrKorab/Arabica/blob/24349afe523730bc0b04e5bfd02af6468174a843/arabica/__init__.py#L1) . On another note, you should make use of `Makefile` to have an automated tests.
