### 2024-03-19 #6186 @PetrKorab
@linuxscout It is easily doable. Should I proceed as suggested  ? :)
