+++ paper.bib
@@ -77,6 +77,7 @@ omi@article{Balduzzi:2023,
     Journal = {ArXiv e-prints},
     Month = may,
     Title = {NLTK: The Natural Language Toolkit},
+    doi = {doi.org/10.48550/arXiv.cs/0205028},
     Year = 2002
 }
 
