+++ paper.bib
@@ -144,6 +144,7 @@ omi@article{Balduzzi:2023,
   	Author = {{Wilkinson}, L.},
   	Booktitle = {The Grammar of Graphics (Statistics and Computing): Second Edition, by Leland Wilkinson.~ISBN 0-387-24544-8.~Published by Springer Science+Business Media, Inc.},
   	Publisher = {Princeton University Press},
+    doi = {doi.org/10.1007/0-387-28695-0},
   	Title = {{The Grammar of Graphics}},
   	Year = 2005
 }
