+++ paper.bib
@@ -1,4 +1,4 @@
-omi@article{Balduzzi:2023,
+@article{Balduzzi:2023,
   	author = {{Balduzzi}, S. and {Rücker}, G. and {Nikolakopoulou}, A. and {Papakonstantinou}, T. and {Salanti}, G. and {Efthimiou}, O. and {Schwarzer}, G.},
     title = {netmeta: An R Package for Network Meta-Analysis Using Frequentist Methods},
   	journal = {Journal of Statistical Software},
