+++ paper.bib
@@ -87,7 +87,7 @@
     Booktitle = {Python for Data Analysis: Second Edition, by Wes McKinney.~ISBN 9781491957660.~Published by O'Reilly Media, Inc.},
     Publisher = {O'Reilly Media, Inc.},
     Title = {Python for data analysis},
-    doi = {doi/10.5555/2458650},
+    doi = {https://dl.acm.org/doi/10.5555/2458650},
     Year = 2013
 }
 
