+++ .travis.yml
@@ -17,7 +17,8 @@ install:
 
 # Run tests
 script:
-  - python paper_graphs.py 
-  - python arabica_freq_unit_test.py
-  - python cappuccino_unit_test.py
-  - python coffee_break_unit_test.py
+  #- python paper_graphs.py 
+  #- python arabica_freq_unit_test.py
+  #- python cappuccino_unit_test.py
+  #- python coffee_break_unit_test.py
+  - python makefile.py
