+++ tests/makefile.py
@@ -1,15 +1,24 @@
-# Makefile for running tests and other tasks
+import subprocess
 
 # Variables
-PYTHON := python3.9
-TEST_FILE := arabica_freq_unit_test.py
+PYTHON = "python3.9"
+TEST_FILE = "test_arabica.py"
 
-.PHONY: test clean
+# Targets
+def test():
+    """Run unit tests."""
+    print("Running tests...")
+    subprocess.run([PYTHON, "-m", "unittest", TEST_FILE])
 
-test:
-	@echo "Running tests..."
-	$(PYTHON) -m unittest $(TEST_FILE)
+def clean():
+    """Clean up generated files."""
+    print("Cleaning up...")
+    # Add commands here to clean up generated files or directories
 
-clean:
-	@echo "Cleaning up..."
-	# Add commands here to clean up generated files or directories
+def lint():
+    """Lint the code."""
+    print("Linting the code...")
+    subprocess.run(["flake8", "."])  # Assuming flake8 is installed
+
+if __name__ == "__main__":
+    test()
