+++ tests/makefile.py
@@ -5,6 +5,10 @@ PYTHON = "python3.9"
 TEST_FILE_1 = "arabica_freq_unit_test.py"
 TEST_FILE_2 = "cappuccino_unit_test.py"
 TEST_FILE_3 = "coffee_break_unit_test.py"
+TEST_FILE_4 = "paper_graphs.py"
+
+
+paper_graphs.py
 
 
 # Targets
@@ -14,6 +18,7 @@ def test():
     subprocess.run([PYTHON, "-m", "unittest", TEST_FILE_1])
     subprocess.run([PYTHON, "-m", "unittest", TEST_FILE_2])
     subprocess.run([PYTHON, "-m", "unittest", TEST_FILE_3])
+    subprocess.run([PYTHON, "-m", "unittest", TEST_FILE_4])
 
 
 def clean():
