### 2024-03-23 #6186 @PetrKorab
@oliviaguest I've updated the setup.py, added the missing DOIs (for the resources where DOI was assigned), and used makefile.py for unit testing, as the referee suggested. Should I make further revisions?
