+++ paper.bib
@@ -77,7 +77,7 @@
     Journal = {ArXiv e-prints},
     Month = may,
     Title = {NLTK: The Natural Language Toolkit},
-    doi = {doi.org/10.48550/arXiv.cs/0205028},
+    doi = {10.48550/arXiv.cs/0205028},
     Year = 2002
 }
 
@@ -87,7 +87,7 @@
     Booktitle = {Python for Data Analysis: Second Edition, by Wes McKinney.~ISBN 9781491957660.~Published by O'Reilly Media, Inc.},
     Publisher = {O'Reilly Media, Inc.},
     Title = {Python for data analysis},
-    doi = {https://dl.acm.org/doi/10.5555/2458650},
+    doi = {10.5555/2458650},
     Year = 2013
 }
 
@@ -145,7 +145,7 @@
   	Author = {{Wilkinson}, L.},
   	Booktitle = {The Grammar of Graphics (Statistics and Computing): Second Edition, by Leland Wilkinson.~ISBN 0-387-24544-8.~Published by Springer Science+Business Media, Inc.},
   	Publisher = {Princeton University Press},
-    doi = {doi.org/10.1007/0-387-28695-0},
+    doi = {10.1007/0-387-28695-0},
   	Title = {{The Grammar of Graphics}},
   	Year = 2005
 }
