### 2024-04-15 #6186 @oliviaguest
@PetrKorab can you tidy up the double brackets in the citations? You can do it by citing without square brackets.
