+++ paper.bib
@@ -1,150 +1,150 @@
-@article{Balduzzi:2023,
-  	author = {{Balduzzi}, S. and {Rücker}, G. and {Nikolakopoulou}, A. and {Papakonstantinou}, T. and {Salanti}, G. and {Efthimiou}, O. and {Schwarzer}, G.},
-    title = {netmeta: An R Package for Network Meta-Analysis Using Frequentist Methods},
-  	journal = {Journal of Statistical Software},
-  	keywords = {Network Meta-Analysis, R, Inconsistency, Component Network Meta-Analysis},
+@article(Balduzzi:2023,
+  	author = (Balduzzi, S. and Rücker, G. and Nikolakopoulou, A. and Papakonstantinou, T. and Salanti, G. and Efthimiou, O. and Schwarzer, G.),
+    title = (netmeta: An R Package for Network Meta-Analysis Using Frequentist Methods),
+  	journal = (Journal of Statistical Software),
+  	keywords = (Network Meta-Analysis, R, Inconsistency, Component Network Meta-Analysis),
     year = 2023,
     volume = 558,
   	month = mar,
-    doi = {10.18637/jss.v106.i02},
-    url = {https://www.jstatsoft.org/article/view/v106i02},
-}
+    doi = (10.18637/jss.v106.i02),
+    url = (https://www.jstatsoft.org/article/view/v106i02),
+)
 
-@misc{cleantext:2021,
-    author = {{Guidiva}, P.},
-    title = {Cleantext—An open-source python package to clean raw text data},
-    year = {2021},
-    publisher = {PyPI},
-    journal = {Python Package Index},
-    url = {https://pypi.org/project/cleantext}
-}
+@misccleantext:2021,
+    author = (Guidiva, P.),
+    title = (Cleantext—An open-source python package to clean raw text data),
+    year = (2021),
+    publisher = (PyPI),
+    journal = (Python Package Index),
+    url = (https://pypi.org/project/cleantext)
+)
 
-@misc{finvader:2023,
-    author = {{Koráb}, P.},
-    title = {FinVADER: VADER sentiment classifier updated with financial lexicons},
-    year = {2023},
-    publisher = {PyPI},
-    journal = {Python Package Index},
-    url = {https://pypi.org/project/finvader}
-}
+@misc(finvader:2023,
+    author = (Koráb, P.),
+    title = (FinVADER: VADER sentiment classifier updated with financial lexicons),
+    year = (2023),
+    publisher = (PyPI),
+    journal = (Python Package Index),
+    url = (https://pypi.org/project/finvader)
+)
 
-@article{Hunter:2007,
-  	author = {{Hunter}, J.~D.},
-    title = {Matplotlib: A 2D Graphics Environment},
-  	journal = {Computing in Science & Engineering},
-  	keywords = {Python, scripting languages, application development, scientific programming},
+@article(Hunter:2007,
+  	author = (Hunter), J.~D.),
+    title = (Matplotlib: A 2D Graphics Environment),
+  	journal = (Computing in Science & Engineering),
+  	keywords = (Python, scripting languages, application development, scientific programming),
     year = 2007,
     volume = 9,
   	month = mar,
-    doi = {10.1109/MCSE.2007.55},
-    url = {https://ieeexplore.ieee.org/document/4160265},
-}
+    doi = (10.1109/MCSE.2007.55),
+    url = (https://ieeexplore.ieee.org/document/4160265),
+)
 
-@article{Hutto:2014,
-    author = {{Hutto}, C. and {Gilbert}, E.},
-    title = {VADER: A Parsimonious Rule-Based Model for Sentiment Analysis of Social Media Text},
[…]
