+++ paper.bib
@@ -73,7 +73,7 @@
 @article(Loper:2002,
     url = (https://arxiv.org/abs/cs/0205028),
     Archiveprefix = (arXiv),
-    Author = (Loper), E. and (Bird), S.),
+    Author = (Loper, E. and Bird, S.),
     Journal = (ArXiv e-prints),
     Month = may,
     Title = (NLTK: The Natural Language Toolkit),
@@ -91,7 +91,7 @@
 )
 
 @article(Mikolajewicz:2019,
-    author = (Mikolajewicz), N. and Komarova, S.~V. ),
+    author = (Mikolajewicz, N. and Komarova, S.~V.),
     title = (Meta-Analytic Methodology for Basic Research: A Practical Guide),
     journal = (Frontiers in Physiology),
     keywords = (meta-analysis, basic research, rapid review, systematic review, MATLAB, methodology),
