+++ paper.bib
@@ -41,7 +41,7 @@
 )
 
 @article(Hutto:2014,
-    author = (Hutto), C. and (Gilbert), E.),
+    author = (Hutto, C. and Gilbert, E.),
     title = (VADER: A Parsimonious Rule-Based Model for Sentiment Analysis of Social Media Text),
     journal = (Proceedings of the International AAAI Conference on Web and Social Media),
     keywords = (Sentiment Analysis, Social Media, Validated Sentiment Lexicon, Human Centered Computing),
@@ -54,7 +54,7 @@
 
 @book(Jenks:1977,
     url = (https://openlibrary.org/books/OL22018775M/Optimal_data_classification_for_choropleth_maps),
-    Author = (Jenks), G.~F..),
+    Author = (Jenks, G.~F..),
     Booktitle = (Optimal data classification for choropleth maps, by  George F. Jenks. Published by University of Kansas),
     Publisher = (University of Kansas),
     Title = (Optimal data classification for choropleth maps),
