+++ paper.bib
@@ -29,7 +29,7 @@
 )
 
 @article(Hunter:2007,
-  	author = (Hunter), J.~D.),
+  	author = (Hunter, J.~D.),
     title = (Matplotlib: A 2D Graphics Environment),
   	journal = (Computing in Science & Engineering),
   	keywords = (Python, scripting languages, application development, scientific programming),
@@ -62,7 +62,7 @@
 )
 
 @misc(jenkspy:2023,
-    author = (Viry), T.),
+    author = (Viry, T.),
     title = (Jenkspy- Compute Natural Breaks (Fisher-Jenks algorithm)),
     year = (2023),
     publisher = (PyPI),
