+++ paper.bib
@@ -10,7 +10,7 @@
     url = (https://www.jstatsoft.org/article/view/v106i02),
 )
 
-@misccleantext:2021,
+@misc(cleantext:2021,
     author = (Guidiva, P.),
     title = (Cleantext—An open-source python package to clean raw text data),
     year = (2021),
@@ -83,7 +83,7 @@
 
 @book(McKinney:2013,
     url = (https://www.oreilly.com/library/view/python-for-data/9781491957653/),
-    Author = (McKinney), W.),
+    Author = (McKinney, W.),
     Booktitle = (Python for Data Analysis: Second Edition, by Wes McKinney.~ISBN 9781491957660.~Published by O'Reilly Media, Inc.),
     Publisher = (O'Reilly Media, Inc.),
     Title = (Python for data analysis),
