+++ paper.bib
@@ -103,7 +103,7 @@
 )
 
 @misc(network:2017,
-    author = (White), I.),
+    author = (White, I.),
     title = (NETWORK: Stata module to perform network meta-analysis),
     year = (2017),
     publisher = (IDEAS/RePEc),
@@ -131,7 +131,7 @@
 )
 
 @misc(textblob:2021,
-    author = (Loria), S.),
+    author = (Loria, S.),
     title = (Textblob - Simple, Pythonic text processing. Sentiment analysis, part-of-speech tagging, noun phrase parsing, and more),
     year = (2021),
     publisher = (PyPI),
