### 2024-04-15 #6186 @PetrKorab
@oliviaguest I've updated the references, but I'm not sure if it the current version is fomally correct. If not, please, send me a reference for citations. Thanks.
