### 2024-04-25 #6186 @oliviaguest
@PetrKorab I didn't mean that and in fact that might be a bug? Please fix that too.

What I mean is, see these:

> Unlike the current packages to perform meta-analysis ((White, 2017); (Mikolajewicz & Komarova, 2019); (Balduzzi et al., 2023)), the package leverages text mining methods for

> While some existing text-processing packages (e.g. (Zerrouki, 2022)) focus on a specific

They do need to have double brackets. I can show you how to fix this if you do not know how, but please fix the compilaton bugs first?
