### 2024-04-29 #6186 @oliviaguest
@PetrKorab OK, I think I will do a PR to explain. Perhaps it's just that you have not seen this type of citation format previously.
