### 2024-04-29 #6186 @oliviaguest
>  Check that the archive title, author list, version tag, and the license are correct
