### 2024-04-29 #6186 @oliviaguest
OK, so once that PR is dealt with can you make sure the version is OK? I see pypi has 1.7.7 but above is 1.7.1, which is right for JOSS? Otherwise, we're good to go. 😊
