### 2024-04-30 #6186 @PetrKorab
@oliviaguest Many thanks for the explanation. Yes, I came from the Economics field where the citation system is a bit different. For JOSS, 1.7.7 is correct. Thanks ! :)
