### 2024-05-02 bd19af0
Added Trigrams line graph for Y
